refactor: add changes based PR #4
