+++ src/pages/background/index.tsx
@@ -1,7 +1,7 @@
 import { BackgroundUI } from "./interface";
 import type { ICard } from "@ptypes/components/card/ICard";
 
-function Background(props: ICard) {
+const Background = (props: ICard) => {
     const { children } = props;
 
     return (
