+++ src/types/components/box/IBox/index.ts
@@ -1,0 +1,15 @@
+interface IBox {
+  children?: React.ReactNode;
+  padding?: string;
+  margin?: string;
+  borderRadius?: string;
+  width?: string;
+  height?: string;
+  appearance?: string;
+  flexDirection?: string;
+  gap?: string;
+  justifyContent?: string;
+  alignItems?: string;
+};
+
+export type{ IBox };
