--- src/types/pages/modal/IModal/index.ts
@@ -1,5 +1,0 @@
-interface IModal {
-    $isMobile: boolean;
-}
-
-export type { IModal };
