+++ src/types/pages/modal/IModalUI/index.ts
@@ -1,6 +1,5 @@
 interface IModalUI {
     width: string;
-    $isMobile?: boolean;
 }
 
 export type { IModalUI };
