+++ styles.css
@@ -1,11 +1,8 @@
 body,
-html {
+html,
+#root {
+  box-sizing: border-box;
   padding: 0;
   margin: 0;
-  height: auto;
-}
-
-body {
-  margin: 0;
-  padding: 0;
+  height: 100vh;
 }
